modularized ontology tool util functions
